fix(runtime): 数値門の写し漏れ・二重門を補完（レビュー L-3/L-5/L-6/L-7）
- attention_qk i8a8 の scale 有限判定を f32 語で見る（兄弟 2 カーネルの Math.fround 門の
  写し漏れ — 公開 API の attrs 門との二重化・現行経路で誤値は出ない）
- embedding params へ i4 の hidden % group_size 門（linear の k % group_size と同型 —
  カーネル直呼び経路の唯一の門。IR 経由は format 層が守る）
- assertEps の下限を f32 最小正規化数 2^-126 へ（subnormal は FTZ ドライバで 0 に潰れ
  「GPU 側だけ eps=0」が復活する — 明示 MUST の未達を閉じる。契約変更なので受理側の
  テスト 1e-45 を拒否側へ移した）
- CPU 参照の eps に Math.fround（スカラ attr は f32 語の規律 — 1 か所だけの非一様。
  最悪相対影響 2.98e-8 で既存 tolerance 帯は不変）

+++ packages/runtime/src/reference/ops.ts
@@ -928,7 +928,8 @@ export const referenceLayerNorm = (
       const deviation = x.data[base + i] - mean;
       squares += deviation * deviation;
     }
-    const inv = 1 / Math.sqrt(squares / dim + eps);
+    // eps は f32 語で運ばれるスカラ attr なので参照側も f32 へ丸めてから使う（他のスカラと同じ規律）
+    const inv = 1 / Math.sqrt(squares / dim + Math.fround(eps));
     for (let i = 0; i < dim; i += 1) {
       out[base + i] = Math.fround((x.data[base + i] - mean) * inv * weight.data[i] + bias.data[i]);
     }
@@ -961,7 +962,8 @@ export const referenceRmsNorm = (
     const base = row * dim;
     let squares = 0;
     for (let i = 0; i < dim; i += 1) squares += x.data[base + i] * x.data[base + i];
-    const inv = 1 / Math.sqrt(squares / dim + eps);
+    // eps は f32 語で運ばれるスカラ attr なので参照側も f32 へ丸めてから使う（layer_norm と同じ）
+    const inv = 1 / Math.sqrt(squares / dim + Math.fround(eps));
     for (let i = 0; i < dim; i += 1) {
       out[base + i] = Math.fround(x.data[base + i] * inv * weight.data[i]);
     }
